Fixed a big issue with loading natively. Turns out in trap_init, the kernel goes and replaces the idtr with a read-only version of the IDT address so that it doesn't leak the kernel's location in memory. Fixed some issues in the gdb script which makes it easier to restore interrupted state. Fixed some other issues. Requires a make detach.

+++ granary/basic_block.cc
@@ -237,7 +237,7 @@ namespace granary {
     {
         // Round our search address down to `BB_INFO_BYTE_ALIGNMENT` before we
         // search for the meta info magic value.
-        uintptr_t addr(reinterpret_cast<uint64_t>(current_pc_));
+        uintptr_t addr(reinterpret_cast<uintptr_t>(current_pc_));
         addr -= (addr % BB_INFO_BYTE_ALIGNMENT);
 
         // Go search for the basic block info by finding potential
